+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+## [6.3.0] - 2026-01-??
+### Changed
+- Add fermata with DocumentBuilder.addAnnotation("articulation", "fermata"), deprecated DocumentBuilder.addFermata().
+- Paint color key "fermata" is deprecated. Fermata uses now "annotation" color key.
+
+### Fixed
+- Fermata lauyout issues.
+
+## Deprecated
+
 ## [6.2.1] - 2026-01-12
 ### Fixed
 - Made SSR-safe for use in Docusaurus website.